requirements.txt
fastapi<0.111,>=0.110
pydantic<2.8,>=2.7
starlette>=0.37,<0.38
httpx>=0.24,<0.27
uvicorn[standard]>=0.29
pytest>=8,<9
passlib>=1.7,<2
psycopg2-binary>=2.9
testing.postgresql>=1.3
